modified some parameters for the SPH algo so that the fluid is more stable in situations where the water elevation  is very high. Also modified the BC method so that it doesn't rely on ObstacleStatic during the BC calculation

+++ Assets/Scripts/SPH/Core/MeshObsCPU.cs
@@ -303,7 +303,7 @@ public class MeshObsCPU : MonoBehaviour
         OP.ObstacleStatic o_static = new OP.ObstacleStatic();
         OP.ObstacleDynamic o_dynamic = new OP.ObstacleDynamic();
         o_static.index = index;
-        o_dynamic.index = index;
+        o_dynamic.index = (uint)index;
         o_dynamic.hasChanged = 1;
 
         // ===== GENERATING VERTICES DATA ===== //
